Refactor aws smithy types (#2638)
## Motivation and Context
This PR refactors `aws-smithy-types` crate. 
`Blob`, `Datetime`, `Number` and `Document` structs now goes to it's own
files.
No changes on feature is introduced.

This is a child-PR of https://github.com/awslabs/smithy-rs/pull/2616 .
However, it is completely independent.

PR that introduces same changes were previously merged to
`unstable-serde` branch, however, it has not been merged to main branch.

## Testing
NA

## Checklist
This PR does not introduce API changes.
----

_By submitting this pull request, I confirm that you can use, modify,
copy, and redistribute this contribution, under the terms of your
choice._
